Added new season, added script to save data for old seasons, saved old season, added links to switch between seasons
