+++ app/data/drivers.ts
@@ -1,0 +1,54 @@
+export const drivers2022 = {
+  erinDrivers: [
+    'Verstappen',
+    'Sainz',
+    'Russell',
+    'Norris',
+    'Gasly',
+    'Alonso',
+    'Magnussen',
+    'Albon',
+    'Bottas',
+    'Zhou',
+  ],
+  andrewDrivers: [
+    'Leclerc',
+    'Hamilton',
+    'Perez',
+    'Ricciardo',
+    'Ocon',
+    'Schumacher',
+    'Tsunoda',
+    'Vettel',
+    'Hulkenberg',
+    'Stroll',
+    'Latifi',
+  ],
+};
+
+export const drivers2023 = {
+  erinDrivers: [
+    'Leclerc',
+    'Perez',
+    'Russell',
+    'Norris',
+    'Zhou',
+    'Alonso',
+    'Hulkenberg',
+    'Albon',
+    'Magnussen',
+    'Tsunoda',
+  ],
+  andrewDrivers: [
+    'Verstappen',
+    'Sainz',
+    'Hamilton',
+    'Gasly',
+    'Ocon',
+    'Piastri',
+    'Bottas',
+    'Stroll',
+    'De Vries',
+    'Sargent',
+  ],
+};
