+++ app/data/index.ts
@@ -1,0 +1,1 @@
+export const defaultSeason = '2023';
