+++ app/routes/index.tsx
@@ -1,91 +1,19 @@
-import {useLoaderData} from 'remix';
-import cheerio from 'cheerio';
+import {LoaderFunction, useLoaderData} from 'remix';
+import {Link} from 'remix';
+import {defaultSeason} from '~/data';
+import {LoaderResponse} from '~/types';
+import {getSeason, get2022Data} from '~/utilities/season.server';
 
-const erinDrivers = [
-  'Verstappen',
-  'Sainz',
-  'Russell',
-  'Norris',
-  'Gasly',
-  'Alonso',
-  'Magnussen',
-  'Albon',
-  'Bottas',
-  'Zhou',
-];
-const andrewDrivers = [
-  'Leclerc',
-  'Hamilton',
-  'Perez',
-  'Ricciardo',
-  'Ocon',
-  'Schumacher',
-  'Tsunoda',
-  'Vettel',
-  'Hulkenberg',
-  'Stroll',
-  'Latifi',
-];
+export const loader: LoaderFunction = async ({request}) => {
+  const url = new URL(request.url);
+  const season = url.searchParams.get('season') || defaultSeason;
 
-interface DriverData {
-  name: string;
-  points: number;
-}
+  if (season === '2022') {
+    return get2022Data();
+  }
 
-interface TeamData {
-  drivers: DriverData[];
-  total: number;
-}
-
-interface LoaderResponse {
-  erin: TeamData;
-  andrew: TeamData;
-}
-
-export async function loader() {
-  const res = await fetch(
-    'https://www.formula1.com/en/results.html/2022/drivers.html'
-  );
-
-  const html = await res.text();
-  const $ = cheerio.load(html);
-  const driverRows = $('.resultsarchive-table tr');
-
-  const erinDriverData: DriverData[] = [];
-  const andrewDriverData: DriverData[] = [];
-
-  driverRows.map((_, node) => {
-    const name = $(node).find('.hide-for-mobile').text();
-    const points = $(node).find('td.dark.bold').text();
-
-    if (isNaN(Number(points))) {
-      return;
-    }
-
-    const driver = {name, points: Number(points)};
-
-    if (erinDrivers.includes(name)) {
-      erinDriverData.push(driver);
-    } else if (andrewDrivers.includes(driver.name)) {
-      andrewDriverData.push(driver);
-    }
-  });
-
-  return {
-    erin: {
-      drivers: erinDriverData,
-      total: erinDriverData.reduce((acc, current) => {
-        return acc + current.points;
-      }, 0),
-    },
-    andrew: {
-      drivers: andrewDriverData,
-      total: andrewDriverData.reduce((acc, current) => {
-        return acc + current.points;
-      }, 0),
-    },
-  };
-}
+  return await getSeason();
+};
 
 export default function Index() {
   const data = useLoaderData<LoaderResponse>();
@@ -93,6 +21,11 @@ export default function Index() {
   return (
     <div>
       <h1>Fantasy Race Cars</h1>
+      <nav>
+        <Link to="/?season=2022">2022</Link>
+        <Link to="/?season=2023">2023</Link>
+      </nav>
+      <h2 className="season-header">Season: 2022</h2>
       <main>
         <div className="contestant">
           <h2>Erin - {data.erin.total}</h2>
