+++ app/styles/global.css
@@ -80,13 +80,30 @@ h2 {
   }
 }
 
+nav {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  gap: 2rem;
+  padding: 1rem 1rem;
+}
+
+nav a {
+  color: var(--main-text-color);
+}
+
+.season-header {
+  text-align: center;
+  margin-bottom: 1rem;
+}
+
 main {
   max-width: 600px;
   margin: 0 auto;
   display: flex;
   justify-content: space-between;
   gap: 2rem;
-  padding: 2rem 1rem;
+  padding: 0rem 1rem;
 }
 
 .contestant {
