+++ app/types.ts
@@ -1,0 +1,14 @@
+export interface DriverData {
+  name: string;
+  points: number;
+}
+
+export interface TeamData {
+  drivers: DriverData[];
+  total: number;
+}
+
+export interface LoaderResponse {
+  erin: TeamData;
+  andrew: TeamData;
+}
